ref: update a methods of class People

+++ src/main/java/com/example/demo/web/controller/PeopleController.java
@@ -30,7 +30,7 @@ public class PeopleController {
     }
 
     @GetMapping("/id/{id}")
-    public ResponseEntity<ResponsePeopleDto> findById(@PathVariable int id){
+    public ResponseEntity<ResponsePeopleDto> findById(@PathVariable Long id){
         People guy = service.findById(id);
         ResponsePeopleDto dto = PeopleMapper.toDto(guy);
         return ResponseEntity.ok().body(dto);
@@ -44,13 +44,13 @@ public class PeopleController {
     }
 
     @PatchMapping("/id/{id}")
-    public ResponseEntity<People> updateName(@PathVariable int id, @RequestBody @Valid RequestNameUpdate nome){
+    public ResponseEntity<People> updateName(@PathVariable Long id, @RequestBody @Valid RequestNameUpdate nome){
         People guy = service.updateName(id, nome);
         return ResponseEntity.ok().body(guy);
     }
 
     @DeleteMapping("/delete/{id}")
-    public ResponseEntity<Void> delebeById(@PathVariable int id){
+    public ResponseEntity<Void> delebeById(@PathVariable Long id){
         service.deleteById(id);
         return ResponseEntity.noContent().build();
     }
